Chest-X-Ray-main/Chest-X-Ray-main/backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
grad-cam==1.5.5
h11==0.16.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy-loader==0.4
markupsafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.1.1
pluggy==1.6.0
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.4
pytest==8.4.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.9.9
torch==2.8.0
torchvision==0.23.0
torchxrayvision==1.3.5
tqdm==4.67.1
ttach==0.0.3
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
